irasaneh/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bleach==4.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Cython==0.29.28
defusedxml==0.7.1
Django==4.0.2
django-braces==1.15.0
django-crispy-forms==1.14.0
django-etc==1.3.1
django-filter==21.1
django-location-field==2.1.0
django-model-utils==4.2.0
django-summernote==0.8.20.0
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.7
python-decouple==3.6
python-gettext==4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tzdata==2021.5
urllib3==1.26.9
webencodings==0.5.1
zipp==3.8.0
